role permission implement done

+++ app/Http/Controllers/Backend/Admin/RoleController.php
@@ -19,25 +19,41 @@ class RoleController extends Controller
      */
     public function index(Request $request)
     {
-        $name = null;
-        $roles = AdminRole::latest();
-        if ($request->has('name') && !empty($request->input('name'))) {
-            $name = $request->name;
-            $roles = $roles->where('name', 'like', '%' . $name . '%');
+        if (auth()->user()->can('role_read')) {
+            $name = null;
+            $roles = AdminRole::latest();
+            if ($request->has('name') && !empty($request->input('name'))) {
+                $name = $request->name;
+                $roles = $roles->where('name', 'like', '%' . $name . '%');
+            }
+
+            $roles = $roles->paginate(15);
+
+
+            $can_read = auth()->user()->can('role_read') ? "" : "style=display:none";
+            $can_update = auth()->user()->can('role_update') ? "" :
+                "style=display:none";
+            $can_delete = auth()->user()->can('role_delete') ? "" :
+                "style=display:none";
+
+            return view('backend.admin.role.index', compact('roles', 'name', 'can_read', 'can_update', 'can_delete'));
+        } else {
+            $link = "admin.dashboard";
+            return view('error.403', compact('link'));
         }
-
-        $roles = $roles->paginate(15);
-
-
-        return view('backend.admin.role.index', compact('roles', 'name'));
     }
     /**
      * Show the form for creating a new resource.
      */
     public function create()
     {
-        $modules = Module::whereIsLabel(0)->whereIsVisibileToRole(1)->whereStatus(1)->get();
-        return view('backend.admin.role.create', compact('modules'));
+        if (auth()->user()->can('role_create')) {
+            $modules = Module::whereIsLabel(0)->whereIsVisibileToRole(1)->whereStatus(1)->get();
+            return view('backend.admin.role.create', compact('modules'));
+        } else {
+            $link = "admin.dashboard";
+            return view('error.403', compact('link'));
+        }
     }
 
     /**
@@ -46,31 +62,35 @@ class RoleController extends Controller
     public function store(Request $request)
     {
         if ($request->ajax()) {
-            $rules = [
-                'name' => 'required|unique:roles,name',
-            ];
+            if (auth()->user()->can('role_create')) {
+                $rules = [
+                    'name' => 'required|unique:roles,name',
+                ];
 
-            $validator = Validator::make($request->all(), $rules);
-            if ($validator->fails()) {
-                return response()->json([
-                    'type' => 'error',
-                    'errors' => $validator->getMessageBag()->toArray()
-                ]);
-            } else {
+                $validator = Validator::make($request->all(), $rules);
+                if ($validator->fails()) {
+                    return response()->json([
+                        'type' => 'error',
+                        'errors' => $validator->getMessageBag()->toArray()
+                    ]);
+                } else {
 
 
-                DB::beginTransaction();
-                try {
-                    $role = Role::create(['name' => $request->input('name')]);
-                    $permissions = array_map('intval', explode(",", $request->input('permissions')));
-                    $role->syncPermissions($permissions);
+                    DB::beginTransaction();
+                    try {
+                        $role = Role::create(['name' => $request->input('name')]);
+                        $permissions = array_map('intval', explode(",", $request->input('permissions')));
+                        $role->syncPermissions($permissions);
 
-                    DB::commit();
-                    return response()->json(['type' => 'success', 'message' => "Successfully Created"]);
-                } catch (Exception $e) {
-                    DB::rollback();
-                    return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
+                        DB::commit();
+                        return response()->json(['type' => 'success', 'message' => "Successfully Created"]);
+                    } catch (Exception $e) {
+                        DB::rollback();
+                        return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
+                    }
                 }
+            } else {
+                abort(403, 'Sorry, you are not authorized to access this page');
             }
         } else {
             return response()->json(['status' => 'false', 'message' => "Access only ajax request"]);
@@ -90,8 +110,13 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
-        $modules = Module::whereIsLabel(0)->whereIsVisibileToRole(1)->get();
-        return view('backend.admin.role.edit', compact('role', 'modules'));
+        if (auth()->user()->can('role_update')) {
+            $modules = Module::whereIsLabel(0)->whereIsVisibileToRole(1)->get();
+            return view('backend.admin.role.edit', compact('role', 'modules'));
+        } else {
+            $link = "admin.dashboard";
+            return view('error.403', compact('link'));
+        }
     }
 
     /**
@@ -100,46 +125,49 @@ class RoleController extends Controller
     public function update(Request $request, Role $role)
     {
         if ($request->ajax()) {
-            // Setup the validator
-            $rules = [
-                'name' => 'required|unique:roles,name,' . $role->id,
-            ];
+            if (auth()->user()->can('role_update')) {
+                $rules = [
+                    'name' => 'required|unique:roles,name,' . $role->id,
+                ];
 
-            $validator = Validator::make($request->all(), $rules);
-            if ($validator->fails()) {
-                return response()->json([
-                    'type' => 'error',
-                    'errors' => $validator->getMessageBag()->toArray()
-                ]);
-            } else {
+                $validator = Validator::make($request->all(), $rules);
+                if ($validator->fails()) {
+                    return response()->json([
+                        'type' => 'error',
+                        'errors' => $validator->getMessageBag()->toArray()
+                    ]);
+                } else {
 
-                DB::beginTransaction();
-                try {
-                    $role->name = $request->input('name');
-                    $role->status = $request->input('status');
-                    $role->save();
+                    DB::beginTransaction();
+                    try {
+                        $role->name = $request->input('name');
+                        $role->status = $request->input('status');
+                        $role->save();
 
-                    $permissions = $request->input('permissions');
+                        $permissions = $request->input('permissions');
 
-                    if (isset($permissions)) {
-                        $permissions = array_map('intval', explode(",", $permissions));
-                        $role->syncPermissions($permissions);  //If one or more role is selected associate user to roles
-                    } else {
-                        //If no role is selected remove exisiting permissions associated to a role
-                        $all_permissions = $role->getAllPermissions();
+                        if (isset($permissions)) {
+                            $permissions = array_map('intval', explode(",", $permissions));
+                            $role->syncPermissions($permissions);  //If one or more role is selected associate user to roles
+                        } else {
+                            //If no role is selected remove exisiting permissions associated to a role
+                            $all_permissions = $role->getAllPermissions();
 
-                        foreach ($all_permissions as $permission) {
-                            $role->revokePermissionTo($permission); //Remove all permissions associated with role
+                            foreach ($all_permissions as $permission) {
+                                $role->revokePermissionTo($permission); //Remove all permissions associated with role
+                            }
                         }
+
+                        DB::commit();
+
+                        return response()->json(['type' => 'success', 'message' => "Successfully Updated"]);
+                    } catch (Exception $e) {
+                        DB::rollback();
+                        return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
                     }
-
-                    DB::commit();
-
-                    return response()->json(['type' => 'success', 'message' => "Successfully Updated"]);
-                } catch (Exception $e) {
-                    DB::rollback();
-                    return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
                 }
+            } else {
+                abort(403, 'Sorry, you are not authorized to access this page');
             }
         } else {
             return response()->json(['status' => 'false', 'message' => "Access only ajax request"]);
@@ -152,15 +180,18 @@ class RoleController extends Controller
     public function destroy(Request $request, AdminRole $role)
     {
         if ($request->ajax()) {
-
-            DB::beginTransaction();
-            try {
-                $role->delete();
-                DB::commit();
-                return response()->json(['type' => 'success', 'message' => 'Successfully Deleted']);
-            } catch (Exception $e) {
-                DB::rollback();
-                return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
+            if (auth()->user()->can('role_delete')) {
+                DB::beginTransaction();
+                try {
+                    $role->delete();
+                    DB::commit();
+                    return response()->json(['type' => 'success', 'message' => 'Successfully Deleted']);
+                } catch (Exception $e) {
+                    DB::rollback();
+                    return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
+                }
+            } else {
+                abort(403, 'Sorry, you are not authorized to access this page');
             }
         } else {
             return response()->json(['status' => 'false', 'message' => "Access only ajax request"]);
@@ -169,35 +200,44 @@ class RoleController extends Controller
 
     public function trashList(Request $request)
     {
-        $name = null;
-        $roles = AdminRole::latest('deleted_at')->onlyTrashed();
-        if ($request->has('name') && !empty($request->input('name'))) {
-            $name = $request->name;
-            $roles = $roles->where('name', 'like', '%' . $name . '%');
-        }
-        $roles = $roles->paginate(15);
+        if (auth()->user()->can('role_trash')) {
+            $name = null;
+            $roles = AdminRole::latest('deleted_at')->onlyTrashed();
+            if ($request->has('name') && !empty($request->input('name'))) {
+                $name = $request->name;
+                $roles = $roles->where('name', 'like', '%' . $name . '%');
+            }
+            $roles = $roles->paginate(15);
 
-        return view('backend.admin.role.trash', compact('roles', 'name'));
+            return view('backend.admin.role.trash', compact('roles', 'name'));
+        } else {
+            $link = "admin.dashboard";
+            return view('error.403', compact('link'));
+        }
     }
 
     public function restore(Request $request, $id)
     {
         if ($request->ajax()) {
-            $role = AdminRole::withTrashed()->findOrFail($id);
+            if (auth()->user()->can('role_trash')) {
+                $role = AdminRole::withTrashed()->findOrFail($id);
 
-            DB::beginTransaction();
-            try {
+                DB::beginTransaction();
+                try {
 
-                if (!empty($role)) {
-                    $role->restore();
+                    if (!empty($role)) {
+                        $role->restore();
+                    }
+
+                    DB::commit();
+
+                    return response()->json(['type' => 'success', 'message' => "Successfully Restored"]);
+                } catch (Exception $e) {
+                    DB::rollback();
+                    return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
                 }
-
-                DB::commit();
-
-                return response()->json(['type' => 'success', 'message' => "Successfully Restored"]);
-            } catch (Exception $e) {
-                DB::rollback();
-                return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
+            } else {
+                abort(403, 'Sorry, you are not authorized to access this page');
             }
         } else {
             return response()->json(['status' => 'false', 'message' => "Access only ajax request"]);
@@ -207,20 +247,23 @@ class RoleController extends Controller
     public function restoreSelected(Request $request, $ids)
     {
         if ($request->ajax()) {
+            if (auth()->user()->can('role_trash')) {
+                DB::beginTransaction();
+                try {
 
-            DB::beginTransaction();
-            try {
+                    $ids = explode(",", $ids);
 
-                $ids = explode(",", $ids);
+                    AdminRole::withTrashed()->whereIn('id', $ids)->restore();
 
-                AdminRole::withTrashed()->whereIn('id', $ids)->restore();
+                    DB::commit();
 
-                DB::commit();
-
-                return response()->json(['type' => 'success', 'message' => "Successfully Restored"]);
-            } catch (Exception $e) {
-                DB::rollback();
-                return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
+                    return response()->json(['type' => 'success', 'message' => "Successfully Restored"]);
+                } catch (Exception $e) {
+                    DB::rollback();
+                    return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
+                }
+            } else {
+                abort(403, 'Sorry, you are not authorized to access this page');
             }
         } else {
             return response()->json(['status' => 'false', 'message' => "Access only ajax request"]);
@@ -230,19 +273,23 @@ class RoleController extends Controller
     public function permanentDelete(Request $request, $id)
     {
         if ($request->ajax()) {
-            $role = AdminRole::withTrashed()->findOrFail($id);
+            if (auth()->user()->can('role_trash')) {
+                $role = AdminRole::withTrashed()->findOrFail($id);
 
-            DB::beginTransaction();
-            try {
+                DB::beginTransaction();
+                try {
 
-                if (!empty($role)) {
-                    $role->forceDelete();
+                    if (!empty($role)) {
+                        $role->forceDelete();
+                    }
+                    DB::commit();
+                    return response()->json(['type' => 'success', 'message' => "Successfully Removed"]);
+                } catch (Exception $e) {
+                    DB::rollback();
+                    return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
                 }
-                DB::commit();
-                return response()->json(['type' => 'success', 'message' => "Successfully Removed"]);
-            } catch (Exception $e) {
-                DB::rollback();
-                return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
+            } else {
+                abort(403, 'Sorry, you are not authorized to access this page');
             }
         } else {
             return response()->json(['status' => 'false', 'message' => "Access only ajax request"]);
@@ -252,20 +299,23 @@ class RoleController extends Controller
     public function permanentDeleteSelected(Request $request, $ids)
     {
         if ($request->ajax()) {
+            if (auth()->user()->can('role_trash')) {
+                DB::beginTransaction();
+                try {
 
-            DB::beginTransaction();
-            try {
+                    $ids = explode(",", $ids);
 
-                $ids = explode(",", $ids);
+                    AdminRole::withTrashed()->whereIn('id', $ids)->forceDelete();
 
-                AdminRole::withTrashed()->whereIn('id', $ids)->forceDelete();
+                    DB::commit();
 
-                DB::commit();
-
-                return response()->json(['type' => 'success', 'message' => "Successfully Removed"]);
-            } catch (Exception $e) {
-                DB::rollback();
-                return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
+                    return response()->json(['type' => 'success', 'message' => "Successfully Removed"]);
+                } catch (Exception $e) {
+                    DB::rollback();
+                    return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
+                }
+            } else {
+                abort(403, 'Sorry, you are not authorized to access this page');
             }
         } else {
             return response()->json(['status' => 'false', 'message' => "Access only ajax request"]);
